Connection and ResultSet refactoring. Some ResultSet rework
Basically, the duplicate code was removed, which was moved to the corresponding base classes.
All the result set common code is in ResultSetBase and MultiResultSetBase. The driver-specific (mysqli, PDO) behavior is moved in the appropriate ResultSetAdapter, which is injected into the base classes.
Some reworking of the ResultSet in such a way that the behavior of fetch methods has become similar to the behavior of the corresponding mysqli and PDO methods.