Percolate class for Manticore PQ feature (#153)
The PQ is a new feature in Manticore that stored queries in a new index type - percolate -  and can test documents if they match the stored queries.

Also added a parameter for Helper::ShowTables to allow performing ``SHOW TABLES LIKE 'filter'`` (this is available in Sphinx too) .  This is also needed to change in the Helper unit test for the show tables test, as for Percolate unit test a different `.conf` should be used to include a ``pq`` index as well.  In Travis when Manticore is tested, it should be used the `manticore.conf` instead of `sphinx.conf`.